add rmd to benchmarks

+++ benches/cli.rs
@@ -1,5 +1,7 @@
 use std::fs;
 use std::process::Command;
+use std::process::Stdio;
+use std::time::Duration;
 
 use assert_fs::prelude::*;
 use assert_fs::TempDir;
@@ -22,6 +24,10 @@ fn rm() -> Command {
 
 fn rmt() -> Command {
     Command::new("rmt")
+}
+
+fn rmd() -> Command {
+    Command::new("rmd")
 }
 
 fn n_files(n: usize) -> TempDir {
@@ -59,11 +65,24 @@ fn n_nested_folder(n: usize) -> TempDir {
     dir
 }
 
+fn installed(name: &str) -> bool {
+    Command::new(name)
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .spawn()
+        .is_ok()
+}
+
 fn bench_dfs_n_files(c: &mut Criterion) {
     let mut group = c.benchmark_group("dfs n files");
+    group
+        .measurement_time(Duration::from_secs(5))
+        .sample_size(50);
+
     let mut rmx = rmx();
     let mut rm = rm();
     let mut rmt = rmt();
+    let mut rmd = rmd();
 
     let dir = n_files(black_box(500));
     group.bench_function("rmx -rf", |b| {
@@ -76,34 +95,56 @@ fn bench_dfs_n_files(c: &mut Criterion) {
         })
     });
 
-    let dir = n_files(black_box(500));
-    group.bench_function("rm -rf", |b| {
-        b.iter(|| {
-            rm.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rm");
-        })
-    });
+    if installed("rm") {
+        let dir = n_files(black_box(500));
+        group.bench_function("rm -rf", |b| {
+            b.iter(|| {
+                rm.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rm");
+            })
+        });
+    }
 
-    let dir = n_files(black_box(500));
-    group.bench_function("rmt -rf", |b| {
-        b.iter(|| {
-            rmt.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rmt");
-        })
-    });
+    if installed("rmt") {
+        let dir = n_files(black_box(500));
+        group.bench_function("rmt -rf", |b| {
+            b.iter(|| {
+                rmt.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmt");
+            })
+        });
+    }
+
+    if installed("rmd") {
+        let dir = n_files(black_box(500));
+        group.bench_function("rmd -rf", |b| {
+            b.iter(|| {
+                rmd.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmd");
+            })
+        });
+    }
 }
 
 fn bench_dfs_n_nested_folders(c: &mut Criterion) {
     let mut group = c.benchmark_group("dfs n nested folders");
+    group
+        .measurement_time(Duration::from_secs(5))
+        .sample_size(50);
+
     let mut rmx = rmx();
     let mut rm = rm();
     let mut rmt = rmt();
+    let mut rmd = rmd();
 
     let dir = n_nested_folder(black_box(100));
     group.bench_function("rmx -rf", |b| {
@@ -116,34 +157,56 @@ fn bench_dfs_n_nested_folders(c: &mut Criterion) {
         })
     });
 
-    let dir = n_nested_folder(black_box(100));
-    group.bench_function("rm -rf", |b| {
-        b.iter(|| {
-            rm.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rm");
-        })
-    });
+    if installed("rm") {
+        let dir = n_nested_folder(black_box(100));
+        group.bench_function("rm -rf", |b| {
+            b.iter(|| {
+                rm.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rm");
+            })
+        });
+    }
 
-    let dir = n_nested_folder(black_box(100));
-    group.bench_function("rmt -rf", |b| {
-        b.iter(|| {
-            rmt.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rmt");
-        })
-    });
+    if installed("rmt") {
+        let dir = n_nested_folder(black_box(100));
+        group.bench_function("rmt -rf", |b| {
+            b.iter(|| {
+                rmt.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmt");
+            })
+        });
+    }
+
+    if installed("rmd") {
+        let dir = n_nested_folder(black_box(100));
+        group.bench_function("rmd -rf", |b| {
+            b.iter(|| {
+                rmd.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmd");
+            })
+        });
+    }
 }
 
 fn bench_dfs_m_folders_n_nested_each(c: &mut Criterion) {
     let mut group = c.benchmark_group("dfs m nested folder n");
+    group
+        .measurement_time(Duration::from_secs(5))
+        .sample_size(50);
+
     let mut rmx = rmx();
     let mut rm = rm();
     let mut rmt = rmt();
+    let mut rmd = rmd();
 
     let dir = m_nested_folder_n(black_box(20), black_box(100));
     group.bench_function("rmx -rf", |b| {
@@ -156,34 +219,56 @@ fn bench_dfs_m_folders_n_nested_each(c: &mut Criterion) {
         })
     });
 
-    let dir = m_nested_folder_n(black_box(20), black_box(100));
-    group.bench_function("rm -rf", |b| {
-        b.iter(|| {
-            rm.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rm");
-        })
-    });
+    if installed("rm") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rm -rf", |b| {
+            b.iter(|| {
+                rm.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rm");
+            })
+        });
+    }
 
-    let dir = m_nested_folder_n(black_box(20), black_box(100));
-    group.bench_function("rmt -rf", |b| {
-        b.iter(|| {
-            rmt.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rmt");
-        })
-    });
+    if installed("rmt") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rmt -rf", |b| {
+            b.iter(|| {
+                rmt.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmt");
+            })
+        });
+    }
+
+    if installed("rmt") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rmd -rf", |b| {
+            b.iter(|| {
+                rmd.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmd");
+            })
+        });
+    }
 }
 
 fn bench_rip_mode(c: &mut Criterion) {
     let mut group = c.benchmark_group("rip mode");
+    group
+        .measurement_time(Duration::from_secs(5))
+        .sample_size(50);
+
     let mut rmx = rmx();
     let mut rm = rm();
     let mut rmt = rmt();
+    let mut rmd = rmd();
 
     let dir = m_nested_folder_n(black_box(20), black_box(100));
     group.bench_function("rmx --rip", |b| {
@@ -195,27 +280,44 @@ fn bench_rip_mode(c: &mut Criterion) {
         })
     });
 
-    let dir = m_nested_folder_n(black_box(20), black_box(100));
-    group.bench_function("rm -rf", |b| {
-        b.iter(|| {
-            rm.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rm");
-        })
-    });
+    if installed("rm") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rm -rf", |b| {
+            b.iter(|| {
+                rm.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rm");
+            })
+        });
+    }
 
-    let dir = m_nested_folder_n(black_box(20), black_box(100));
-    group.bench_function("rmt -rf", |b| {
-        b.iter(|| {
-            rmt.arg("-r")
-                .arg("-f")
-                .arg(dir.path())
-                .output()
-                .expect("to execute rmt");
-        })
-    });
+    if installed("rmt") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rmt -rf", |b| {
+            b.iter(|| {
+                rmt.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmt");
+            })
+        });
+    }
+
+    if installed("rmt") {
+        let dir = m_nested_folder_n(black_box(20), black_box(100));
+        group.bench_function("rmd -rf", |b| {
+            b.iter(|| {
+                rmd.arg("-r")
+                    .arg("-f")
+                    .arg(dir.path())
+                    .output()
+                    .expect("to execute rmd");
+            })
+        });
+    }
 }
 
 criterion_group!(
